dingux-sdl: Unify emulator menu buttons for the GCW Zero.
The Start+Select key combinations and flicking the Power slider up are now accepted as ways to enter the menu.

gui.cpp: Do not rely on a KEYUP event occurring after the menu is entered, which resets g_key, in the event that the key used to enter the menu is released before the menu appears. This is required for the Power flick, above.